+++ src/functions/tooltip-functions.js
@@ -60,3 +60,51 @@ function returnDomainFromUrl(url, firstLetterIsCapital = true) {
         return '';
     }
 }
+
+function checkTooltipForCollidingWithSideEdges() {
+    if (configs.debugMode)
+        console.log('Checking Selecton tooltip to colliding with side edges...');
+
+    var dx = parseInt(tooltip.style.left.replaceAll('px', ''));
+
+    var tooltipWidth = 12.0;
+    tooltip.querySelectorAll('.selection-popup-button').forEach(function (el) {
+        tooltipWidth += el.offsetWidth;
+    });
+
+    /// Tooltip is off-screen on the left
+    if (dx < 0) {
+
+        if (configs.debugMode)
+            console.log('Tooltip is colliding with left edge. Fixing...');
+
+        tooltip.style.left = '5px';
+
+        /// Shift the arrow to match new position
+        var newLeftPercentForArrow = ((dx * -1) + 5) / tooltipWidth * 100;
+        if (arrow !== null && arrow !== undefined)
+            arrow.style.left = `${50 - newLeftPercentForArrow}%`;
+
+    } else {
+        var screenWidth = window.innerWidth
+            || document.documentElement.clientWidth
+            || document.body.clientWidth;
+
+        var offscreenAmount = (dx + tooltipWidth) - screenWidth + 10;
+
+        /// Tooltip is off-screen on the right
+        if (offscreenAmount > 0) {
+            if (configs.debugMode)
+                console.log('Tooltip is colliding with right edge. Fixing...');
+
+            tooltip.style.left = `${dx - offscreenAmount - 5}px`;
+
+            /// Shift the arrow to match new position
+            var newLeftPercentForArrow = (dx - (dx - offscreenAmount - 5)) / tooltipWidth * 100;
+            arrow.style.left = `${50 + newLeftPercentForArrow}%`;
+        } else {
+            if (configs.debugMode)
+                console.log('Tooltip is not colliding with side edges');
+        }
+    }
+}